devmate factor our test_torch tests
prompt "ok great now can you split test\_torch.py into more smaller
pieces just like you did with test\_basic\_vital\_signs.py?"

[ghstack-poisoned]
